Update GetHTML.bas
improved accuracy when user has used parameters in "job search" - this causes linkedIN to display "This job won't be displayed anymore" although it IS still displayed. 

Code update adds searching in the html for that phrase "again." and if it appears next to the hyperlink that would otherwise be a job we would go check out, it ignores that job instead. 

This doesn't fix everything. behavior is still buggy.